+++ src/define.ts
@@ -32,6 +32,10 @@ export type ConfigSyncMap = {
 export enum Platform {
 	OSX, Windows, Unknown
 }
+export enum OpenEgretServerType {
+	auto = "auto",
+	alert = "alert"
+}
 export interface ConfigObj extends vscode.WorkspaceConfiguration {
 	/**插件是否可用*/
 	enable: boolean;
@@ -39,18 +43,16 @@ export interface ConfigObj extends vscode.WorkspaceConfiguration {
 	devlog: boolean;
 	/**是否弹出错误弹窗*/
 	alertErrorWin: boolean;
-	/**egret项目文件egretProperties.json文件路径(相对于项目根目录)*/
-	egretPropertiesPath: string;
 	/**egret资源配置default.res.json的路径(相对于项目根目录)*/
 	egretResourceJsonPath: string;
 	/**要同步到default.res.json资源文件夹路径(相对于项目根目录)*/
 	egretResourcePath: string;
-	/**exml搜索路径(glob字符串格式)*/
-	exmlSearchGlob: string;
 	/**是否使用外部编辑器打开exml(目前仅支持在ts文件快捷键打开)*/
 	exmlOpenExternal: boolean;
 	/**同步default.res.json的资源*/
 	resMap: ConfigSyncMap,
 	/**同步default.res.json忽略的资源*/
 	resMapIgnore: string[],
+	/**打开项目时怎么开启Egret服务器*/
+	openEgretServer: OpenEgretServerType,
 }