+++ src/egret-server/EgretServer.ts
@@ -5,6 +5,8 @@ import EgretBuild from './EgretBuild';
 import EgretResSync from './EgretResSync';
 import EgretServerBar from './EgretServerBar';
 import EgretService from './EgretService';
+import * as helper from "../helper";
+import { OpenEgretServerType } from '../define';
 
 export default class EgretServer extends Listener {
     private _service: EgretService;
@@ -49,8 +51,21 @@ export default class EgretServer extends Listener {
             this.logger.devlog("constructor egret-helper.egretResSync");
             this._resSync.start()
         }));
-
-        this._service.start();
+        let openEgretType = helper.getConfigObj().openEgretServer;
+        switch (openEgretType) {
+            case OpenEgretServerType.auto:
+                this._service.start();
+                break;
+            case OpenEgretServerType.alert:
+                let menus = [{ title: "确定", v: "ok" }, { title: "取消", v: "cancel" }]
+                vscode.window.showInformationMessage("是否启动Egret服务器?", ...menus).then(result => {
+                    if (!this._service || this._service.isDestroy) return;
+                    if (result && result.v == "ok") {
+                        this._service.start();
+                    }
+                });
+                break;
+        }
         // this._resSync.start();//刚初始化时同步一次
     }
     public get service() {