+++ src/exml/ExmlPathAutoCompleteProvider.ts
@@ -24,9 +24,7 @@ export default class ExmlPathAutoCompleteProvider implements vscode.CompletionIt
 		}
 		workPath = this.normalPath(workPath);
 		if (!workPath) return;
-		let configs = helper.getConfigObj();
-		if (!configs.exmlSearchGlob) return;
-		return vscode.workspace.findFiles(configs.exmlSearchGlob, undefined, undefined, token).then(result => {
+		return vscode.workspace.findFiles("**/*.exml", undefined, undefined, token).then(result => {
 			return result.map(item => {
 				let paths = this.normalPath(item.fsPath).slice(workPath.length);
 				return new CompletionItem(`"${paths}"`)